[ANDROSDK-1229-4] Refactor unit tests
