+++ core/src/test/java/org/hisp/dhis/android/core/imports/internal/conflicts/NonUniqueAttributeConflictShould.kt
@@ -28,30 +28,12 @@
 package org.hisp.dhis.android.core.imports.internal.conflicts
 
 import com.nhaarman.mockitokotlin2.doReturn
-import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
-import org.hisp.dhis.android.core.arch.db.stores.internal.IdentifiableObjectStore
-import org.hisp.dhis.android.core.trackedentity.TrackedEntityAttribute
-import org.junit.Before
 import org.junit.Test
 
-class NonUniqueAttributeConflictShould {
-
-    private val attributeUid = "DI85uC13Bzo"
-
-    private val context: TrackerImportConflictItemContext = mock()
-
-    private val attributeStore: IdentifiableObjectStore<TrackedEntityAttribute> = mock()
-
-    private val attribute: TrackedEntityAttribute = mock()
+internal class NonUniqueAttributeConflictShould : BaseConflictShould() {
 
     private val importConflict = TrackedImportConflictSamples.nonUniqueAttribute(attributeUid, "a value")
-
-    @Before
-    fun setUp() {
-        whenever(context.attributeStore) doReturn attributeStore
-        whenever(attributeStore.selectByUid(attributeUid)) doReturn attribute
-    }
 
     @Test
     fun `Should match error message`() {
